Separate colourlists command file
